+++ tbsim/interventions/README.md
@@ -2,10 +2,12 @@
 
 This folder contains all TBsim intervention modules.
 
+- `dx_products.py` -- TB diagnostic products (`Dx`, `xpert`, `oral_swab`, `fujilam`, `cad_cxr`)
+- `tx_products.py` -- TB treatment products (`Tx`, `dots`, `dots_improved`, `first_line`, `second_line`)
+- `diagnostics.py` -- Diagnostic delivery intervention (`DxDelivery`)
+- `treatments.py` -- Treatment delivery intervention (`TxDelivery`)
 - `bcg.py` -- BCG vaccination product and routine delivery
 - `tpt.py` -- TB preventive therapy products and delivery
-- `tb_treatment.py` -- TB treatment interventions
-- `tb_diagnostic.py` -- TB diagnostic interventions
 - `tb_health_seeking.py` -- Health-seeking behavior
 - `immigration.py` -- Immigration interventions
 - `beta.py` -- Time-varying transmission modifier
