+++ tbsim/interventions/__init__.py
@@ -9,10 +9,8 @@ from .interventions import *
 from .beta import *
 from .tpt import *
 from .bcg import *
-from .tb_treatment import *
 from .tb_drug_types import *
 from .tb_health_seeking import *
-from .tb_diagnostic import *
 from .dx_products import *
 from .tx_products import *
 from .diagnostics import *
